@@ -1,10 +1,9 @@
-'use strict';
-import {describe, it, beforeEach, afterEach, expect, jest} from '@jest/globals';
-import {shallowMount, enableAutoUnmount} from '@vue/test-utils';
-import type {VueWrapper} from '@vue/test-utils';
-import JIcon, {plugin} from '../src';
-import icons, {alert, code, x} from './icon';
-import type {Icon} from '../index';
+import { describe, it, beforeEach, afterEach, expect, vitest } from 'vitest';
+import { shallowMount, enableAutoUnmount } from '@vue/test-utils';
+import type { VueWrapper } from '@vue/test-utils';
+import JIcon, { plugin } from '../src';
+import icons, { alert, code, x } from './icon';
+import type { Icon } from '../index';
 
 enableAutoUnmount(afterEach);
 
@@ -13,16 +12,18 @@ describe('Vue j-icon component', () => {
         describe('register global component', () => {
             it('default name', () => {
                 const element = shallowMount(JIcon, {
-                    props: {icon: alert},
-                    global: {plugins: [[plugin]]}});
+                    props: { icon: alert },
+                    global: { plugins: [ [ plugin ] ] }
+                });
                 expect(element.vm.$.appContext.app.component(JIcon.displayName!)).not.toBeUndefined();
             });
 
             it('custom name', () => {
                 const name = 'CustomName'
                     , element = shallowMount(JIcon, {
-                    props: {icon: alert},
-                    global: {plugins: [[plugin, {name}]]}});
+                    props: { icon: alert },
+                    global: { plugins: [ [ plugin, { name } ] ] }
+                });
                 expect(element.vm.$.appContext.app.component(name)).not.toBeUndefined();
             });
         });
@@ -30,8 +31,8 @@ describe('Vue j-icon component', () => {
         describe('prefix option', () => {
             it('no prefix', () => {
                 const element = shallowMount(JIcon, {
-                    props: {icon: alert},
-                    global: {plugins: [[plugin, {prefix: false}]]}
+                    props: { icon: alert },
+                    global: { plugins: [ [ plugin, { prefix: false } ] ] }
                 });
 
                 expect(element.exists()).toBe(true);
@@ -43,31 +44,31 @@ describe('Vue j-icon component', () => {
 
             it('default prefix', () => {
                 const element = shallowMount(JIcon, {
-                    props: {icon: alert},
-                    global: {plugins: [[plugin, {prefix: true}]]}
+                    props: { icon: alert },
+                    global: { plugins: [ [ plugin, { prefix: true } ] ] }
                 });
 
                 expect(element.exists()).toBe(true);
-                expect(element.classes()).toContain(`icon-${alert.name}`);
+                expect(element.classes()).toContain(`icon-${ alert.name }`);
             });
 
             it('custom prefix', () => {
                 const prefix = 'custom-'
                     , element = shallowMount(JIcon, {
-                    props: {icon: alert},
-                    global: {plugins: [[plugin, {prefix}]]}
+                    props: { icon: alert },
+                    global: { plugins: [ [ plugin, { prefix } ] ] }
                 });
 
                 expect(element.exists()).toBe(true);
-                expect(element.classes()).toContain(`${prefix}${alert.name}`);
+                expect(element.classes()).toContain(`${ prefix }${ alert.name }`);
             })
         });
 
         it('classes option', () => {
-            const value = ['0123456789', '9876543210']
+            const value = [ '0123456789', '9876543210' ]
                 , element = shallowMount(JIcon, {
-                props: {icon: alert},
-                global: {plugins: [[plugin, {classes: [value]}]]}
+                props: { icon: alert },
+                global: { plugins: [ [ plugin, { classes: [ value ] } ] ] }
             });
 
             expect(element.exists()).toBe(true);
@@ -85,22 +86,25 @@ describe('Vue j-icon component', () => {
         }
 
         it('from SVG name', () => {
-            const element = shallowMount(JIcon, {props: {icon: 'x'}, global: {plugins: [[plugin, {icons}]]}});
+            const element = shallowMount(JIcon, {
+                props: { icon: 'x' },
+                global: { plugins: [ [ plugin, { icons } ] ] }
+            });
             assertion(element, x);
         });
 
         it('from icon detail', () => {
-            const element = shallowMount(JIcon, {props: {icon: code}});
+            const element = shallowMount(JIcon, { props: { icon: code } });
             assertion(element, code);
         });
 
         it('size rollback to the viewBox', () => {
-            const element = shallowMount(JIcon, {props: {icon: alert}});
+            const element = shallowMount(JIcon, { props: { icon: alert } });
             assertion(element, alert);
         });
 
         it('inherit undeclared properties', () => {
-            const element = shallowMount(JIcon, {props: {icon: alert, version: '1.1'}});
+            const element = shallowMount(JIcon, { props: { icon: alert, version: '1.1' } });
             assertion(element, alert);
             expect(element.attributes('version')).toBe('1.1');
         });
@@ -109,7 +113,7 @@ describe('Vue j-icon component', () => {
     describe('component props', () => {
         let element: VueWrapper;
         beforeEach(() => {
-            element = shallowMount(JIcon, {props: {icon: x}});
+            element = shallowMount(JIcon, { props: { icon: x } });
             expect(element.exists()).toBeTruthy();
         });
 
@@ -117,7 +121,7 @@ describe('Vue j-icon component', () => {
             expect(element.attributes('width')).toBe(String(x.size![0]));
 
             const width = String(x.size![0] + 1);
-            await element.setProps({width});
+            await element.setProps({ width });
 
             expect(element.attributes('width')).toBe(width);
         });
@@ -126,33 +130,33 @@ describe('Vue j-icon component', () => {
             expect(element.attributes('height')).toBe(String(x.size![1]));
 
             const height = String(x.size![1] + 1);
-            await element.setProps({height});
+            await element.setProps({ height });
 
             expect(element.attributes('height')).toBe(height);
         });
 
-        it.each(['0', '1'])('tabindex GE 0, aria-hidden become false', async (value: string) => {
+        it.each([ '0', '1' ])('tabindex GE 0, aria-hidden become false', async (value: string) => {
             expect(element.attributes('aria-hidden')).toBe('true');
 
-            await element.setProps({tabindex: value});
+            await element.setProps({ tabindex: value });
             expect(element.attributes('tabindex')).toBe(value);
             expect(element.attributes('aria-hidden')).toBe('false');
         });
 
         it('tabindex less 0, aria-hidden become true', async () => {
             const value = '-1';
-            element = shallowMount(JIcon, {props: {icon: x, tabindex: value}});
+            element = shallowMount(JIcon, { props: { icon: x, tabindex: value } });
 
             expect(element.attributes('tabindex')).toBe(value);
             expect(element.attributes('aria-hidden')).toBe('true');
         });
 
-        it.each(['invalid', NaN])(
+        it.each([ 'invalid', NaN ])(
             'tabindex invalid string or NaN, not change aria-hidden',
             async (value: number | string) => {
                 expect(element.attributes('aria-hidden')).toBe('true');
 
-                await element.setProps({tabindex: value});
+                await element.setProps({ tabindex: value });
 
                 expect(element.attributes('tabindex')).toBeUndefined();
                 expect(element.attributes('aria-hidden')).toBe('true');
@@ -163,7 +167,7 @@ describe('Vue j-icon component', () => {
             expect(element.attributes('aria-hidden')).toBe('true');
 
             const text = 'aria label';
-            await element.setProps({ariaLabel: text});
+            await element.setProps({ ariaLabel: text });
 
             expect(element.attributes('aria-label')).toBe(text);
             expect(element.attributes('aria-hidden')).toBe('false');
@@ -174,7 +178,7 @@ describe('Vue j-icon component', () => {
             expect(element.attributes('aria-hidden')).toBe('true');
 
             const text = 'aria description';
-            await element.setProps({ariaDescription: text});
+            await element.setProps({ ariaDescription: text });
 
             expect(element.attributes('aria-description')).toBe(text);
             expect(element.attributes('aria-hidden')).toBe('false');
@@ -185,7 +189,7 @@ describe('Vue j-icon component', () => {
             expect(element.attributes('aria-hidden')).toBe('true');
 
             const text = 'title';
-            await element.setProps({title: text});
+            await element.setProps({ title: text });
 
             const titleElement = element.find('title');
 
@@ -200,7 +204,7 @@ describe('Vue j-icon component', () => {
             expect(element.attributes('aria-hidden')).toBe('true');
 
             const text = 'desc';
-            await element.setProps({desc: text});
+            await element.setProps({ desc: text });
 
             const descElement = element.find('desc');
 
@@ -213,8 +217,8 @@ describe('Vue j-icon component', () => {
 
     describe('event', () => {
         it('bind event', () => {
-            const onClick = jest.fn<(event: MouseEvent) => void>()
-                , element = shallowMount(JIcon, {props: {icon: alert}, attrs: {onClick}});
+            const onClick = vitest.fn<[ MouseEvent ], void>()
+                , element = shallowMount(JIcon, { props: { icon: alert }, attrs: { onClick } });
             element.trigger('click');
 
             expect(onClick).toBeCalledTimes(1);
